+++ Foundation/Auth/integrations/release/CheckSystemShape.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Avax\Auth\Integrations\Release;
 
+use FilesystemIterator;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use SplFileInfo;
@@ -76,7 +77,7 @@ final readonly class CheckSystemShape
         $forbiddenDirectories = [];
 
         $iterator = new RecursiveIteratorIterator(
-            iterator: new RecursiveDirectoryIterator(directory: $systemRoot, flags: RecursiveDirectoryIterator::SKIP_DOTS),
+            iterator: new RecursiveDirectoryIterator(directory: $systemRoot, flags: FilesystemIterator::SKIP_DOTS),
             mode    : RecursiveIteratorIterator::SELF_FIRST
         );